# aptos-api/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Aptos Dashboard API",
    "version": "1.0.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health Check",
        "responses": {
          "200": {
            "description": "Healthy"
          }
        }
      }
    },
    "/api/resource": {
      "get": {
        "summary": "Rate Limited Resource",
        "responses": {
          "200": {
            "description": "Successful response"
          }
        }
      }
    },
    "/aptos/providers": {
      "get": {
        "summary": "Get all providers",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "provider": { "type": "string" },
                      "total_tvl": { "type": "number", "format": "float" },
                      "total_volume_1d": { "type": "number", "format": "float" },
                      "total_volume_7d": { "type": "number", "format": "float" },
                      "total_volume_30d": { "type": "number", "format": "float" },
                      "avg_fees_1d": { "type": "number", "format": "float" },
                      "avg_fees_7d": { "type": "number", "format": "float" },
                      "avg_fees_30d": { "type": "number", "format": "float" },
                      "total_pools": { "type": "integer" },
                      "avg_tvl_per_pool": { "type": "number", "format": "float" },
                      "avg_slippage_1d": { "type": "number", "format": "float" },
                      "avg_slippage_7d": { "type": "number", "format": "float" },
                      "avg_slippage_30d": { "type": "number", "format": "float" }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/aptos/providers/{provider_name}": {
      "get": {
        "summary": "Get provider by name",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/aptos/pool/{pool_address}/current": {
      "get": {
        "summary": "Get current pool data",
        "parameters": [
          {
            "name": "pool_address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "404": {
            "description": "Pool not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/aptos/pools": {
      "get": {
        "summary": "Get all current pools",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/aptos/pool/{pool_address}/history": {
      "get": {
        "summary": "Get pool history",
        "parameters": [
          {
            "name": "pool_address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "range",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["day", "week", "month"],
              "default": "day"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid time range"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/top": {
      "get": {
        "summary": "Get top 10 pools",
        "parameters": [
          {
            "name": "metric",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["tvl", "volume_day", "fees_day"],
              "default": "tvl"
            },
            "description": "The metric to sort by. Can be 'tvl', 'volume_day', or 'fees_day'. Defaults to 'tvl'."
          },
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "The provider to filter by. If not provided, no filtering is applied."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid metric"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  }
}